# docs/openapi.generated.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/groups": {
      "get": {
        "operationId": "GroupsController_list",
        "parameters": [
          {
            "name": "mine",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Liste des groupes (?mine=true)",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crée un groupe (staff)",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}": {
      "get": {
        "operationId": "GroupsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail groupe",
        "tags": [
          "groups"
        ]
      },
      "patch": {
        "operationId": "GroupsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Met à jour un groupe (leader ou staff)",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/join": {
      "post": {
        "operationId": "GroupsController_join",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rejoint le groupe",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/leave": {
      "delete": {
        "operationId": "GroupsController_leave",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Quitte le groupe",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/messages": {
      "get": {
        "operationId": "GroupsController_messages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Messages du groupe",
        "tags": [
          "groups"
        ]
      },
      "post": {
        "operationId": "GroupsController_sendMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Envoie un message (diffusé en WS)",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/groups/{id}/read": {
      "post": {
        "operationId": "GroupsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marque les messages comme lus",
        "tags": [
          "groups"
        ]
      }
    },
    "/api/v1/prayers": {
      "get": {
        "operationId": "PrayersController_wall",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mur de prière (public)",
        "tags": [
          "prayers"
        ]
      },
      "post": {
        "operationId": "PrayersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrayerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soumet une prière",
        "tags": [
          "prayers"
        ]
      }
    },
    "/api/v1/prayers/queue": {
      "get": {
        "operationId": "PrayersController_queue",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Champ:direction, ex. createdAt:desc",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Recherche plein texte",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "File de care (staff)",
        "tags": [
          "prayers"
        ]
      }
    },
    "/api/v1/prayers/{id}": {
      "get": {
        "operationId": "PrayersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail prière",
        "tags": [
          "prayers"
        ]
      }
    },
    "/api/v1/prayers/{id}/amen": {
      "post": {
        "operationId": "PrayersController_amen",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Toggle « prayed »",
        "tags": [
          "prayers"
        ]
      }
    },
    "/api/v1/prayers/{id}/status": {
      "patch": {
        "operationId": "PrayersController_setStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrayerStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change le statut (staff)",
        "tags": [
          "prayers"
        ]
      }
    },
    "/api/v1/prayers/{id}/respond": {
      "post": {
        "operationId": "PrayersController_respond",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondPrayerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Répond à une prière (staff)",
        "tags": [
          "prayers"
        ]
      }
    },
...
    "/api/v1/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mes notifications",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/read": {
      "post": {
        "operationId": "NotificationsController_markRead",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Marque des notifications comme lues",
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/v1/notifications/broadcast": {
      "post": {
        "operationId": "NotificationsController_broadcast",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Diffuse un push (staff)",
        "tags": [
          "notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "IPCK API",
    "description": "API de la plateforme IPCK (mobile + dashboard)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ]
      },
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "leaderId": {
            "type": "string"
          },
          "meets": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "leaderId": {
            "type": "string"
          },
          "meets": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        }
      },
      "CreatePrayerDto": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "public",
              "anon",
              "private"
            ]
          }
        },
        "required": [
          "text",
          "visibility"
        ]
      },
      "UpdatePrayerStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "answered",
              "rejected"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
      "RespondPrayerDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "MarkReadDto": {
        "type": "object",
        "properties": {
          "ids": {
            "description": "IDs à marquer lus (tous si vide)",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "BroadcastDto": {
        "type": "object",
        "properties": {
          "audience": {
            "type": "string",
            "example": "all",
            "description": "all | devo-subscribers"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "body"
        ]
      }
    }
  }
}